goci-lint: deduplicate code in intergration test
No-op refactoring to move the duplicated code into a shared util file.
