+++ goci-lint/tests/gofmt_test.go
@@ -3,7 +3,6 @@ package goci_test
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"regexp"
 	"testing"
 
@@ -130,17 +129,4 @@ func TestGofmt(t *testing.T) {
 			}
 		})
 	}
-}
-
-func replaceInFile(path, old, new string, once bool) error {
-	data, err := ioutil.ReadFile(path)
-	if err != nil {
-		return err
-	}
-	if once {
-		data = bytes.Replace(data, []byte(old), []byte(new), 1)
-	} else {
-		data = bytes.ReplaceAll(data, []byte(old), []byte(new))
-	}
-	return ioutil.WriteFile(path, data, 0o666)
 }
