+++ goci-lint/tests/goimports_test.go
@@ -3,7 +3,6 @@ package goci_test
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"strings"
 	"testing"
 
@@ -184,17 +183,4 @@ func TestGofmt(t *testing.T) {
 			}
 		})
 	}
-}
-
-func replaceInFile(path, old, new string, once bool) error {
-	data, err := ioutil.ReadFile(path)
-	if err != nil {
-		return err
-	}
-	if once {
-		data = bytes.Replace(data, []byte(old), []byte(new), 1)
-	} else {
-		data = bytes.ReplaceAll(data, []byte(old), []byte(new))
-	}
-	return ioutil.WriteFile(path, data, 0o666)
 }
