+++ goci-lint/tests/util.go
@@ -1,0 +1,19 @@
+package goci_test
+
+import (
+	"bytes"
+	"io/ioutil"
+)
+
+func replaceInFile(path, old, new string, once bool) error {
+	data, err := ioutil.ReadFile(path)
+	if err != nil {
+		return err
+	}
+	if once {
+		data = bytes.Replace(data, []byte(old), []byte(new), 1)
+	} else {
+		data = bytes.ReplaceAll(data, []byte(old), []byte(new))
+	}
+	return ioutil.WriteFile(path, data, 0o666)
+}
